requirements.txt
aiohttp
python-socketio
aiortc